tests/wpt/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-source-media.html

<!doctype html>
<title>the &lt;source> media attribute has no effect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<video><source src="resources/delayed-broken-video.py" media="none"></video>
<script>
test(function() {
  var v = document.querySelector('video');
  var s = document.querySelector('source');
  assert_equals(v.networkState, v.NETWORK_LOADING);
  assert_equals(v.currentSrc, s.src);
});
</script>
