Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-source-media-expected.txt


FAIL the <source> media attribute has no effect assert_equals: expected 2 but got 3
